refactor: file DB 구조 변경에 따른 Entity 대응
