+++ src/main/java/org/depromeet/sambad/moyeo/file/application/FileService.java
@@ -3,7 +3,6 @@ package org.depromeet.sambad.moyeo.file.application;
 import lombok.RequiredArgsConstructor;
 import org.depromeet.sambad.moyeo.file.domain.FileEntity;
 import org.depromeet.sambad.moyeo.file.domain.FileRepository;
-import org.depromeet.sambad.moyeo.file.domain.FileUploader;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -19,7 +18,7 @@ public class FileService {
 		// TODO: add url validation
 		try {
 			String uploadedPath = fileUploader.upload(fileUrl);
-			FileEntity fileEntity = FileEntity.of(uploadedPath, fileUploader.getFileNameFromUrl(fileUrl));
+			FileEntity fileEntity = FileEntity.of(fileUploader.getFileNameFromUrl(fileUrl), uploadedPath);
 			return fileRepository.save(fileEntity);
 		} catch (IOException e) {
 			throw new RuntimeException("Failed to upload file", e);
